workers/requirements.txt
moviepy==1.0.3
pillow==10.1.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
elevenlabs==0.2.26
openai==1.3.5
